refactor: extract trace store import path helpers
